+++ demo/nextjs/pages/full-screen-exit-button/index.tsx
@@ -1,0 +1,57 @@
+import * as React from 'react';
+import { PrimaryButton, Viewer, Worker } from '@react-pdf-viewer/core';
+import { fullScreenPlugin } from '@react-pdf-viewer/full-screen';
+
+const IndexPage = () => {
+    const fullScreenPluginInstance = fullScreenPlugin({
+        renderExitFullScreenButton: (props) => (
+            <div
+                style={{
+                    bottom: '1rem',
+                    position: 'fixed',
+                    right: '1rem',
+                }}
+            >
+                <PrimaryButton onClick={props.onClick}>Exit</PrimaryButton>
+            </div>
+        ),
+    });
+    const { EnterFullScreenButton } = fullScreenPluginInstance;
+
+    return (
+        <Worker workerUrl="https://unpkg.com/pdfjs-dist@2.10.377/build/pdf.worker.js">
+            <div
+                style={{
+                    border: '1px solid rgba(0, 0, 0, 0.1)',
+                    display: 'flex',
+                    flexDirection: 'column',
+                    height: '50rem',
+                    margin: '5rem auto',
+                    width: '64rem',
+                }}
+            >
+                <div
+                    style={{
+                        alignItems: 'center',
+                        backgroundColor: '#eeeeee',
+                        borderBottom: '1px solid rgba(0, 0, 0, 0.1)',
+                        display: 'flex',
+                        padding: '4px',
+                    }}
+                >
+                    <EnterFullScreenButton />
+                </div>
+                <div
+                    style={{
+                        flex: 1,
+                        overflow: 'hidden',
+                    }}
+                >
+                    <Viewer fileUrl="/pdf-open-parameters.pdf" plugins={[fullScreenPluginInstance]} />
+                </div>
+            </div>
+        </Worker>
+    );
+};
+
+export default IndexPage;